Add token checking when deleting blogs

+++ osa4/bloglist/controllers/blogs.js
@@ -34,6 +34,17 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const userId = decodedToken.id
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() !== userId ) {
+    return response.status(401).json({ error: `only the blog owner can delete this blog` })
+  }
+  
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
